moved MAIN from products to api
